refactor: je passe tout dans utils.py
